Handle nil body in govspeak_with_attachments_to_html
I'm about to move the content for a WorldwideOrganisation from the
CorporateInformationPage about page onto the WorldwideOrganisation
itself. Once that move is completed, I'll need to call
GovspeakHelper#govspeak_with_attachments_to_html with the body text to
render the page. The body attribute is not required and so it can be
blank.

Prior to this commit #govspeak_with_attachments_to_html raised an
exception if body was nil. And the `if` statement in
PublishingApi::WorkingGroupPresenter#body was protecting against this.
It doesn't feel ideal adding `if` statements everywhere the body
attribute is used so I first considered trying to set the default value
to be "" in the database, but MySQL text columns don't support a default
value. I then tried using
ActiveRecord::Attributes::ClassMethods#attribute to set a default value,
but that didn't appear to work, presumably because `body` is a
translated attribute.

So the least worst option seems to be to have
GovspeakHelper#govspeak_with_attachments_to_html treat a nil body as if
it were an empty string.

Making this change means we can simplify
PublishingApi::WorkingGroupPresenter#body and we won't need to add
similar `if` statements for places that use the soon to be added
WorldwideOrganisation#body.