Remove redundant assertions from WO steps
WorldwideOrganisation#description & #services were concatenated into the
body of the about page in this commit [1] as part of this PR [2].

The worldwide_organisations_translations.services database column was
belatedly removed in this commit [3] as part of this PR [4].

These assertions have not been doing anything useful since this commit
[5] as part of this PR [6] where the last steps that actually set
translation["description"] or translation["services"] were removed.
Since then the assertions have been checking the content for text values
of nil which I assume are converted into empty strings which in turn
means the assertions always succeed.

Since I'm about to make changes in this area, now seems like a good time
to remedy the situation.

[1]: https://github.com/alphagov/whitehall/commit/b12e7bf5cbc6161e9cb28810373c204767be2a54
[2]: https://github.com/alphagov/whitehall/pull/1476
[3]: https://github.com/alphagov/whitehall/commit/65c3b99f1a86c171ad7455fc78ee9a6f93ec14dd
[4]: https://github.com/alphagov/whitehall/pull/7612
[5]: https://github.com/alphagov/whitehall/commit/50256fb046b2762807749c3da184735e210ed9dc
[6]: https://github.com/alphagov/whitehall/pull/1476