Move attachments from about page to WorldwideOrganisation
Trello: https://trello.com/c/hwo2GEzk

This includes the Attachable concern into WorldwideOrganisation,
associates the attachments from the WorldwideOrganisation about us page
with the WorldwideOrganisation itself, and updates the publishing UI and
the public-facing pages accordingly.

In the migration, attachments associated with WorldwideOrganisation
about us pages are updated to be associated with the
WorldwideOrganisations themselves. This is achieved by changing the
attachable which corresponds to the attachable_type & attachable_id in
the attachments table.

I did consider copying the attachments rather than updating their
attachable association, but this would have meant their URL path would
have changed, because it includes the AttachmentData#id. Although it
seems a bit unlikely that the attachment URL paths are out in the wild,
it seems better to keep them working if we can.

I chose to make this a standard migration vs a data migration, because
there's not much data to update so it doesn't take long to run and
because I think it would be challenging to choreograph the deployment of
the code changes with the running of a data migration.

I've attempted to make the migration as independent as possible from the
application code so that it will continue to work at least for a while
irrespective of whether we remove some of the code relating to the about
us page.

It was awkward to make the down method work and so I've left it blank.
If we do need to reverse the change, it shouldn't be too hard to
implement as a separate migration.