Don't republish WO about page when other Corp Info pages change
The Organisation about page is a separate publicly-facing page; whereas
the WorldwideOrganisation about page is not - it's just a container for
content for the WorldwideOrganisation page itself. While at the moment
the WorldwideOrganisation about page has a separate content item, it's
not currently used for rendering and should not exist. So there's no
point in republishing it from this callback.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>