Exclude WO about pages from list of CorporateInfoPages
And explain how the content for the WorldwideOrganisation page can now
be edited.

We should be able to remove the about page filtering once the
WorldwideOrganisation about pages have been removed. However, it might
make sense to retain the explanatory message for a while longer to help
users make the transition.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>