Update people show page to use summary list component
This uses the summary list component rather than edit and delete buttons
and the biography.

It adds rows for title, image and biography and only renders the row
if a value is present.

When an image is put into a summary list value it causes the keys value
to render on at the bottom of the row. I think this is due to the line
being stretched to the width of the image. I've added styling to
ensure the text renders inline with the top of the image.